backend/requirements.txt
flask>=3.0
flask-cors>=4.0
pytest>=8.0
python-dotenv>=1.0
